+++ backend/src/main/java/nl/itvitae/specops/users/UserService.java
@@ -12,7 +12,7 @@ public class UserService {
   private final UserRepository userRepository;
 
   public User save(String username, String password, String employeeName) {
-    return userRepository.save(new User(username, passwordEncoder.encode(password), employeeName));
+    return userRepository.save(new User(username, passwordEncoder.encode(password)));
   }
 
   public User getByName(String name) {
