+++ frontend/src/contexts/SecurityContext.tsx
@@ -1,4 +1,6 @@
 import { FC, ReactNode, createContext, useState } from "react";
+import api from "../services/api-client";
+import axios from "axios";
 
 type ContextType = {
   user: string;
@@ -12,6 +14,11 @@ type ContextType = {
 
   errMsg: string;
   setErrMsg: (error: string) => void;
+
+  success: boolean;
+  setSuccess: (boolean: boolean) => void;
+
+  handleAddUser: (e: React.FormEvent<HTMLFormElement>) => void;
 };
 
 type ProviderType = FC<{ children: ReactNode }>;
@@ -28,6 +35,11 @@ export const SecurityContext = createContext<ContextType>({
 
   errMsg: "",
   setErrMsg: () => {},
+
+  success: false,
+  setSuccess: () => {},
+
+  handleAddUser: () => {},
 });
 
 export const SecurityProvider: ProviderType = ({ children }) => {
@@ -38,7 +50,17 @@ export const SecurityProvider: ProviderType = ({ children }) => {
   const [matchPwd, setMatchPwd] = useState("");
 
   const [errMsg, setErrMsg] = useState("");
+
   const [success, setSuccess] = useState(false);
+
+  const handleAddUser = async (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    try {
+      const response = await axios.post()
+    } catch (err) {
+
+    }
+  };
 
   return (
     <SecurityContext.Provider
@@ -51,6 +73,9 @@ export const SecurityProvider: ProviderType = ({ children }) => {
         setMatchPwd,
         errMsg,
         setErrMsg,
+        success,
+        setSuccess,
+        handleAddUser,
       }}
     >
       {children}
